SolverParameters to RangeActionsOptimizationParameters.LinearOptimizationSolver
